backend/requirements.txt
flask
flask-cors
gunicorn
google-auth
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
flask-compress
google-auth
gunicorn